REFACTOR: 확장성 있는 컴포넌트 구현 고민
- fontSize, isBig, isFilled props를 전달하여 조건부 스타일링 구현
